python2_requirements.txt
absl-py==0.7.1
astor==0.8.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bayesian-optimization==1.0.1
blis==0.2.4
boto==2.49.0
boto3==1.9.199
botocore==1.12.199
bz2file==0.98
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
cycler==0.10.0
cymem==2.0.2
docutils==0.14
enum34==1.1.6
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.17.1
futures==3.3.0
gast==0.2.2
gensim==3.8.0
google-pasta==0.1.7
grpcio==1.22.0
gym==0.14.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mock==3.0.5
murmurhash==1.0.2
numpy==1.16.1
opt-einsum==2.3.2
pandas==0.24.2
pandas-datareader==0.7.4
pathlib==1.0.1
Pillow==6.1.0
plac==0.9.6
preshed==2.0.1
protobuf==3.9.0
psycopg2-binary==2.8.3
pyglet==1.3.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.20.4
scipy==1.2.2
six==1.12.0
sklearn==0.0
smart-open==1.8.4
spacy==2.1.6
SQLAlchemy==1.3.6
srsly==0.0.7
subprocess32==3.5.4
tb-nightly==1.15.0a20190731
tensorflow-estimator-2.0-preview==1.14.0.dev2019073000
tensorflow-hub==0.5.0
termcolor==1.1.0
tf-nightly-2.0-preview==2.0.0.dev20190731
thinc==7.0.8
tqdm==4.32.2
urllib3==1.25.3
wasabi==0.2.2
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
